- setSeed and getSeed function are added to the API function.
- SEED is removed from the selections list, so it is no longer possible to set and get its value using setValue and getValue

- the overridden setValue function in the Gillespie integrator is now deprecated since there is no need to set its seed value independently

- setSeed and getSeed function in hte rrc_api.cpp are now deprecated

- the bug with getting RANDOM_SEED value in Random class is fixed again

+++ source/llvm/Random.cpp
@@ -59,7 +59,7 @@ static Function* createGlobalMappingFunction(const char* funcName,
 
 static int64_t defaultSeed()
 {
-    int64_t seed = Config::getValue(Config::RANDOM_SEED).getAs<int>();
+    int64_t seed = Config::getValue(Config::RANDOM_SEED).getAs<long int>();
     if (seed < 0)
     {
         // system time in mirsoseconds since 1970
